Namespaced, adding StringStore for arguments
ArgOpts class renamed Parser, and put inside namespace ArgOpts.
This is to allow forward declarations of classes, which cannot
be done for nested classes.

StringStore converts to/from strings by overloading the cast
operator. This should allow the expected type of an argument
to be known, and a useful error message printed.